data_generator/image_data/images_robotics/src/requirements.txt
pillow
random2
matplotlib
